backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
pymongo==4.12.0
python-dotenv==1.1.0
twilio==9.5.2
PyJWT==2.10.1 